Using Interfaces to Describe Data
