+++ src/app/album.ts
@@ -1,0 +1,25 @@
+import { Track } from "./track";
+
+export interface Album {
+    name:string;
+    releaseDate:string;
+    coverImage:string;
+    tracks: Track
+}
+
+/* 
+export class _albumUrl{
+
+
+    private id:number;
+    private artist:string;
+    private album:album[];
+  
+  }
+  
+  export class album{
+
+  
+  }
+  
+ */
