+++ src/app/product.service.ts
@@ -1,7 +1,9 @@
 import { Injectable } from '@angular/core';
 import {Http, Response} from '@angular/http';
 import 'rxjs/add/operator/map';
+import { Observable } from 'rxjs/Observable';
 import { map } from 'rxjs/operators';
+import { Album } from './album';
 
 @Injectable()
 export class ProductService {
@@ -9,34 +11,9 @@ export class ProductService {
   private _albumUrl="../assets/album.json";
   constructor(private _http: Http) { }
 
-  getAlbum(id:number){
-   return this._http.get(this._albumUrl).pipe(map(response => response.json()));
+  getAlbum(id:number):Observable<Album>{
+   return this._http.get(this._albumUrl).pipe(map(response => <Album>response.json()));
   }
 
-
-}
-
-export class _albumUrl{
-
-
-  private id:number;
-  private artist:string;
-  private album:album[];
-
-}
-
-export class album{
-  name:string;
-  releaseDate:string;
-  coverImage:string;
-  tracks: tracks[];
-
-}
-
-export class tracks{
-  trackNumber:number;
-  trackName:string;
-  trackLength:string;
-  trackPrice:number;
 
 }
