+++ src/app/track.ts
@@ -1,0 +1,9 @@
+export interface Track {
+        
+        trackNumber:number;
+        trackName:string;
+        trackLength:string;
+        trackPrice:number;
+      
+      
+}
